Make window less monolithic
